Fix random number generator constants.
This is probably a bug in the source parser.   One of the old constants was
greater than the value that could be placed in a SmallInt.   However, it was
still forced into a SmallInt.   When the image was written out, the value
was mangled.  Or perhaps when it was read back in.  Either way, it was not
good.

This commit changes the constants to ones that are valid SmallInt values.